api/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.1
django-admin-sortable2==2.1.3
django-environ==0.9.0
django-filter==22.1
django-model-utils==4.2.0
django-nested-admin==3.4.0
django-ory-auth==0.1.3
django-stubs==1.12.0
django-stubs-ext==0.5.0
djangorestframework==3.13.1
djangorestframework-stubs==1.7.0
drf-flex-fields==1.0.0
drf-nested-routers==0.93.4
gevent==21.12.0
greenlet==1.1.3
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
mypy==0.971
mypy-extensions==0.4.3
psycopg2-binary==2.9.3
python-monkey-business==1.0.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
types-Markdown==3.4.1
types-pytz==2022.2.1.0
types-PyYAML==6.0.11
types-requests==2.28.9
types-urllib3==1.26.23
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
zope.event==4.5.0
zope.interface==5.4.0
